# Monotonic entropy source