Fixes #6
Supersedes #15 

PTAL @peterbourgon @bruth @aybabtme