@tsenart 2018-09-28
@peterbourgon: I can't reproduce that performance degradation. Here are my results:

```
name         old time/op    new time/op    delta
Timestamp-8    0.31ns ± 4%    0.32ns ± 2%   ~     (p=1.000 n=10+10)

name         old speed      new speed      delta
Timestamp-8  25.5GB/s ± 4%  25.3GB/s ± 2%   ~     (p=0.436 n=10+10)
```