@tsenart 2018-09-28
I have one concern with this change set and the parent ULID repo that specifies it. Please chime in @alizain @ulid.

When using cryptographic entropy, users have the need and expectation that the entropy will **always** be truly random and of crypto grade quality. But this change set totally throws that out the door, introducing deterministic entropy within the same millisecond. In fact, I think the change is not OK even with non-crypto grade randomness.

If an attacker can within one millisecond learn about the last ULID generated and predict potential successive ULIDs within that millisecond, well, that's an information leak, and it may well be used in nefarious ways I'm not able to think of right now.

First and foremost, I'd like your take on whether this is a real problem or if I am over-estimating the issue.

Second, if it is an issue, we need to think of an alternative that solves it and to treat it with the sensitivity that may or may not be warranted.

Let me know your thoughts.