@peterbourgon 2018-09-28
IMO that (valid!) concern can be accommodated with a large disclaimer in the doc comment for func Monotonic.