@tsenart 2018-09-28
> it comes with the landscape if one wants a within-some-period-of-time monotonic source that the output will become predictable.

Yeah, I am wondering if this is a necessity though. Could there be a non-deterministic but efficient way to generate a string of entropy bytes that is strictly lexicographically larger than another string of bytes of the same length?

I think I may have stumbled into an idea that could work better than plain increments.